s/alert/alert_with_message/, resolve WSL name conflict
Looks like the version of Ubuntu that comes with Windows Subsystem for
Linux (WSL) has a naming conflict with the `alert` function in my .bash_profile.
It appears to overlap somehow with something in `libnotify-bin`.

Not really using this anywhere actively so just changing to a more
verbose `alert_with_message`.

```
mrundle$ uname -r
6.6.87.2-microsoft-standard-WSL2
mrundle$ alert
Command 'notify-send' not found, but can be installed with:
sudo apt install libnotify-bin
```