windows: add PS1 visual cue for wsl vs. git-bash
WSL is the preferred experience, but when navigating VSCode it runs as a separate
process. This doesn't have any meaningful effect except when you want
to open a file in your current VSCode session; running `code` in a
VSCode WSL terminal results in a new window popping up.

So my current strategy for this is to just add `wsl`/`git-bash` tags
to the PS1 prompt.